Email verification (#19)
* Created account creation page and added barebones fields (email, password + password confirm) as placeholders

* Added basic registration page as a base, only has
fields for email, password, and password confirm.
Working on proper firebase connection.

* Added register button on account_creation_view to finalize registration (does nothing yet)

* Firebase initialization, committing here incase I need to rollback to the first step of this process.

* Testing authentication emulation as well as passing the email and password through to the authentication function.

* Email and password are now stored on Firebase! Issue: Credential says it's not being used? (line 41) So, will need to see why that is.

* Passwords can now be checked with one another, added alerts to confirm if registration was successful or not.

* Commented out main

* Commented out main for my page to remove conflicts
with main.dart

* Cleaned up some old comments, working towards getting the proper error message to show up when creating an account.

* HUGE code cleanup for account creation. Moved functions below the class, reformated the widget to match the style of the login page. As of this commit, confirm password is still using the old style, but this is for reference as i move code over.

* Password confirmation working, just needs to update as entered. Right now, the formatting for the page is messed up, the error codes are very basic, and overall, I should convert the if statement to a switch statement.

* Can now show/hide password (hidden by default) for registration screen. Corrected the textboxes to fit the style of the other pages of the app.

* Password strength requirements now appear as you type! Code is still fairly messy- plenty of debug checks, and a load of if statements.

* Password strength check and confirm password check are now fully implemented!

* Small update: Added first and last name onto registration page.

* Information is now stored in collections through the registration page!

* Working on registration error message being shown as an alert. Examples: If the email already exists, a popup will notify the user, asking them to use another email, or if they would like a password reset sent to that email.

* Basic alerts setup to notify user of a successful registration or if passwords do not match. Was attempting to get the button to be hidden when all fields were blank, but that was proving difficult, so will be skipping that and instead opting for a "please fill out all fields" error message.

* Updated alert messages to catch more errors. Annoyingly, can't seem to get it to catch email is in use despite the code Google gave on it's website of password authentication skipping RIGHT over it's own try/catch block!

* Alerts for the registration page now properly work! My issue was that I completely forgot try/catch blocks will always execute the final code regardless if an exception is caught or not, allowing me to pass the email in use error along with getting the alerts to pop up properly!

* Successful registration now routes the user to the home page. Started to add more detailed comments before functions (see: emailCheckAndStore) will make comments more akin to this in the future.

* Added show/hide password button for login page, changed font sizes to be more readable for presentation purposes.

* Was unable to get the email validation to work- currently commented out. Login now redirects to homepage for valid email and password, still need to change how the message displays however.

* Adding user to collections has been changed in order to sync with the upload file function. Thank you Maaz for this change!

* Moved account_creation_view into auth section while also shortening it's code, adding that code into account_creation_control. Email verification now works properly!

* Changed registeraccount -> registerView on the login page. Forgot to push this with my last push + pubspec files

* Missed a work in progress import- now gone!

* Forgot password functionality implemented

* Fixed auth issue

* Added label texts to sign in page for email and password and updated the button on the registration page for a consistent style.

---------
