+++ lib/src/auth/account_creation_view.dart
@@ -1,14 +1,16 @@
-// import 'package:firebase_auth/firebase_auth.dart';
+// ignore_for_file: prefer_const_constructors
+
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
-// import 'package:microbial_source_tracking/src/home/home_view.dart';
-import 'package:microbial_source_tracking/src/auth/account_creation_view.dart';
-import 'package:microbial_source_tracking/src/widgets/login_controller.dart';
+import 'package:microbial_source_tracking/src/login/login_view.dart';
+import '../widgets/account_creation_controller.dart';
 
-import '../auth/password_reset_view.dart';
-
-class LoginView extends StatelessWidget {
-  const LoginView({super.key});
+//RegisterView is the name of this widget,
+//refer to RegisterView for routing purposes
+class RegisterView extends StatelessWidget {
+  const RegisterView({
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -27,7 +29,7 @@ class LoginView extends StatelessWidget {
                       child: SvgPicture.asset('assets/images/glwa_logo.svg')),
                   const SizedBox(height: 50),
                   const Text(
-                    'Log in to your account',
+                    'Create a new account',
                     style: TextStyle(
                         color: Colors.black,
                         fontSize: 20,
@@ -35,32 +37,9 @@ class LoginView extends StatelessWidget {
                   ),
                   const SizedBox(height: 15),
 
-                  // Login widget
-                  Login(),
-
+                  //Register widget
+                  Register(),
                   const SizedBox(height: 20),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(horizontal: 30),
-                    child: Row(
-                        mainAxisAlignment: MainAxisAlignment.end,
-                        children: [
-                          InkWell(
-                            onTap: () {
-                              Navigator.push(
-                                context,
-                                MaterialPageRoute(
-                                  builder: (context) =>
-                                      const PasswordResetView(),
-                                ),
-                              );
-                            },
-                            child: const Text('Forgot password?',
-                                style: TextStyle(
-                                    color: Colors.black, fontSize: 20)),
-                          )
-                        ]),
-                  ),
-                  const SizedBox(height: 10),
                 ],
               )),
         ),
@@ -69,19 +48,19 @@ class LoginView extends StatelessWidget {
         height: 50,
         alignment: Alignment.topCenter,
         child: Row(mainAxisAlignment: MainAxisAlignment.center, children: [
-          const Text('Don\'t have an account? ',
+          const Text('Already have an account? ',
               style: TextStyle(fontSize: 20)),
           InkWell(
             onTap: () {
               Navigator.push(
                 context,
                 MaterialPageRoute(
-                  builder: (context) => const RegisterView(),
+                  builder: (context) => const LoginView(),
                 ),
               );
             },
             child: const Text(
-              'Sign up here',
+              'Return to login',
               style: TextStyle(
                   color: Colors.blue,
                   fontWeight: FontWeight.bold,
